+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 2.0.26
 ++++++
 * Allow VSTS as a remote source location
+* Add 'acr import' command.
 
 2.0.25
 ++++++